REFACT/WIP: Change how styles combined
Because of various issues (*), I disable the priority notion. Instead all
matching styles are applied.
In order to avoid to apply stuff where they should not, zero-width patterns
shall be used. See the various functions altered in this commit.

However, it's likelly we need to say: "DO NOT change that". Further
modifications are required. All the default styles also need to be fixe.